@@ -4,23 +4,25 @@ namespace GerardRoche\TravisCITesting;
 
 use PDO;
 
-class MySQLTest extends \PHPUnit_Framework_TestCase
+class MySQLTest extends \PHPUnit\Framework\TestCase
 {
     protected static $pdo;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         if (self::$pdo === null) {
-            self::$pdo = new PDO(
+            $pdo = new PDO(
                 'mysql:charset=utf8',
                 DBUNIT_USERNAME,
                 DBUNIT_PASSWORD
             );
 
-            self::$pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-            self::$pdo->setAttribute(PDO::ATTR_EMULATE_PREPARES, false);
-            self::$pdo->setAttribute(PDO::ATTR_STRINGIFY_FETCHES, false);
-            self::$pdo->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
+            $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+            $pdo->setAttribute(PDO::ATTR_EMULATE_PREPARES, false);
+            $pdo->setAttribute(PDO::ATTR_STRINGIFY_FETCHES, false);
+            $pdo->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
+
+            self::$pdo = $pdo;
         }
     }
 
