@@ -2,7 +2,7 @@
 
 namespace GerardRoche\TravisCITesting;
 
-class PHPIniTest extends \PHPUnit_Framework_TestCase
+class PHPIniTest extends \PHPUnit\Framework\TestCase
 {
     public function testErrorReporting()
     {
